Updated SNP call scripts.
30th July 2024

I went through the scripts in 03_processing/02_original_sample_sheet to call
SNPs. I am aiming to be agnostic about the actual identity of the genotypes,
and will add scripts to validate in a separate folder.

Note that scripts 5, 6, 9, 10 and 11 currently don't work, and probably don't
need to, and that the sample names in the output VCF file are base file names.

To do:
    - Rename this something like "02_snp_calls"
    - Remove the SNPmatch stuff and move these to a folder for validating genotypes.
    - Either fix, or remove the base calibration stuff.
    - Add a script to rename the samples to give plate and positions.
        This will be enough to validate genotypes. Give the samples proper names
        only after validation.

 Changes to be committed:
	modified:   03_processing/02_original_sample_sheet/05_base_recalibration.sh
	modified:   03_processing/02_original_sample_sheet/06_rename_bams.R
	modified:   03_processing/02_original_sample_sheet/07_call_SNPs.sh
	modified:   03_processing/02_original_sample_sheet/08_merge_VCF.sh
	new file:   03_processing/02_original_sample_sheet/call_SNPs_GATK.sh
	new file:   03_processing/02_original_sample_sheet/readme.md