Attempt at validating genotypes
Commit of a first attempt to construct a pipeline to validate genotypes visually

This relies on 02_library/check_trio.py that calculates similarity between and
offspring and two candidate parents across their genomes. This divides the genome
into windows of roughly equal number of SNPs and estimates pi between the
offsping and each candidate parent in each window, and between the two parents.
It returns a dataframe with a row for each window, giving pi between each pair
of individuals.

The presentation file gives an overview of the results

This did not really work.
Pi is way too high to be credible.
I think I should restrict this to SNPs that differ between the parents.

 Changes to be committed:
	new file:   02_library/check_trio.py
	new file:   02_library/rank_candidate_parents.py
	new file:   02_library/vcf_to_HDF5.py
	new file:   03_processing/05_quality_control/00_create_test_HDF5.sh
	new file:   03_processing/05_quality_control/01_create_HDF5.sh
	new file:   03_processing/05_quality_control/02_rank_candidates.sh
	new file:   03_processing/05_quality_control/03_check_trios.sh
	new file:   03_processing/05_quality_control/04_random_trio.sh
	new file:   03_processing/05_quality_control/05_plot_trios.R
	new file:   03_processing/05_quality_control/plan_for_qc.md
	new file:   07_presentations/240117_can_we_use_these_data/240117_can_we_use_these_data.Rmd