Updates to SNP calls for F8s, first round.
Changed column names in sample sheet from barcode{1,2} to index{2,1}
because methlab expects the latter.
Changed renaming python script to match.
(notice order change - for some reason the order of adapters in fastq
filenames is opposite to what is published in the Nextera adapter list).

Added a line to the unzipping script to move the unzipped fastq files
to a new directory, because scratch-cbe seems to delete them after a day
otherwise.