Scripts to pre-process SNPs for the parental lines.
SNP genotypes for 1163 inbred lines are in
01_data/03_parental_genotypes/1163g.179kB.prior15.gauss4.ts99.5.BIALLELIC.vcf.gz
This includes all but two of the parental lines.
Accessions 1137 and 1074 are missing, and I sequenced them seprately.

These scripts calculate summary statistics on those SNPs, then filter for minor
allele count, missingness, quality, and depth.

This also creates a 'targets file' of variable SNP positions to call for cross
data later.

 Changes to be committed:
	new file:   03_processing/01_parental_SNP_matrix/01_inspect_initial_snp_matrix.sh
	new file:   03_processing/01_parental_SNP_matrix/02_plot_inspection.R
	new file:   03_processing/01_parental_SNP_matrix/03_filter_VCF.sh
	new file:   03_processing/01_parental_SNP_matrix/04_inspect_filtered_snp_matrix.sh
	new file:   03_processing/01_parental_SNP_matrix/05_plot_filtered_vcf.R
	new file:   03_processing/01_parental_SNP_matrix/readme.md