Attempt at expected decay in D.
I attempted to compare the observed decay in D to that expected given
parental LD based on

D_0 * (1-c)(1-2c)^n

where D_0 is the initial LD in the parents
c is recombination fraction
and n is the number of generations of selfing.

I do not think I have implemented this correctly, because the expected
values look far too high. I should check Falconer and MacKay when I am
back in my office and check again.