Added functional scripts to prepare plink files for observed datasets for simulating phenotypes, run those simulations using GCTA, and perform GWAS on each in GEMMA.
Many of the phenotypes end up having quite large (negative) values, such
that they overlap with -9, which Plink and Gemma (sometimes?) use to mean
missing data. I think this is fine because I don'think GCTA outputs any
actual missing data, and if it does they are unlikely to be *exactly* -9.

An additional issue is that the simulations currently generate 3200
phenotype files (4 values for nloci, 4 heritabilities, 50 replicates of
each is 800; this is done separately for four input datasets). It seems
that SLURM won't let you submit job arrays with indices above 3000, so
I had to find an adhoc work around to get it to work.

One solution would be to run jobs for the four datasets for each parameter
setting together in one script, which would cut the number of jobs by 4.

 Changes to be committed:
	new file:   05_results/22_simulated_datasets/03_create_plink.sh
	new file:   05_results/22_simulated_datasets/04_simulate_phenotypes.sh
	new file:   05_results/22_simulated_datasets/05_run_gemma.sh
 Changes not staged for commit:
	modified:   05_results/03_long_range_ld/02_LD_between_windows.py
	modified:   05_results/03_long_range_ld/03_find_SNPs_in_LD.sh
	modified:   05_results/03_long_range_ld/04_plot_heatmap.R
	modified:   05_results/22_simulated_datasets/02_permute_parents.sh