LD between chromosomes without minimum threshold
In 02_LD_between_windows.py I removed the sum between
masks for missing data because it is impossible to ensure that windows are always
the same shape, this was throwing exceptions.

Fixed header in the output file of 03_find_SNPs_in_LD.sh

Edited 04_plot_heatmap.R to handle updated header in its input.
The input file is now massive, so this is now read in chunks, and filtered for
rows with r2 above a threshold on the fly.
Removed the histograms from this script

Added 05_plot_histograms.R to plot histograms of R2 and D between chromosomes.

 Changes to be committed:
   (use "git reset HEAD <file>..." to unstage)

	modified:   05_results/03_long_range_ld/02_LD_between_windows.py
	modified:   05_results/03_long_range_ld/03_find_SNPs_in_LD.sh
	modified:   05_results/03_long_range_ld/04_plot_heatmap.R
	new file:   05_results/03_long_range_ld/05_plot_histograms.R