Revised filtering of global SNP panels
Revised filtering of main global VCF file with 1163 accessions to use only
biallelic SNPs with limited heterozygosity. The RegMap panel is then filtered
to match the same positions (which leads to a massive drop in the number of
markers in that panel).

Updated paths in subsequent processing scripts that use those output files.

 Changes to be committed:
	modified:   03_processing/01_parental_SNP_matrix/01_inspect_initial_snp_matrix.sh
	modified:   03_processing/01_parental_SNP_matrix/03_filter_VCF.sh
	modified:   03_processing/01_parental_SNP_matrix/06_process_RegMap.sh